requirements.txt
# Core dependencies
telethon>=1.28.5
python-dotenv>=1.0.0
pandas>=2.0.0
tqdm>=4.65.0

# Database
psycopg2-binary>=2.9.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter notebook
jupyter>=1.0.0
ipykernel>=6.25.0

# CLI and utilities
typer>=0.9.0
rich>=13.0.0

# Date/time handling
pytz>=2023.3 